+++ src/features/XIT/CMDL/CommandLists.vue
@@ -1,29 +1,38 @@
 <script setup lang="ts">
 import { showTileOverlay, showConfirmationOverlay } from '@src/infrastructure/prun-ui/tile-overlay';
-import CreateNoteOverlay from '@src/features/XIT/NOTE/CreateNoteOverlay.vue';
+import CreateCommandListOverlay from './CreateCommandListOverlay.vue';
 import PrunButton from '@src/components/PrunButton.vue';
 import ActionBar from '@src/components/ActionBar.vue';
 import { showBuffer } from '@src/infrastructure/prun-ui/buffers';
 import { userData } from '@src/store/user-data';
-import { createNote, deleteNote } from '@src/store/notes';
 import { vDraggable } from 'vue-draggable-plus';
 import grip from '@src/utils/grip.module.css';
 import fa from '@src/utils/font-awesome.module.css';
 import PrunLink from '@src/components/PrunLink.vue';
+import { createId } from '@src/store/create-id';
 
-function createNewNote(ev: Event) {
-  showTileOverlay(ev, CreateNoteOverlay, {
+function createNew(ev: Event) {
+  showTileOverlay(ev, CreateCommandListOverlay, {
     onCreate: (name: string) => {
-      const id = createNote(name);
-      return showBuffer(`XIT NOTE ${id}`);
+      const id = createId();
+      userData.commandLists.push({
+        id,
+        name,
+        commands: [],
+      });
+      return showBuffer(`XIT CMDL ${id.substring(0, 8)}`);
     },
   });
 }
 
-function confirmDelete(ev: Event, note: UserData.Note) {
-  showConfirmationOverlay(ev, () => deleteNote(note), {
-    message: `Are you sure you want to delete the note "${note.name}"?`,
-  });
+function confirmDelete(ev: Event, list: UserData.CommandList) {
+  showConfirmationOverlay(
+    ev,
+    () => (userData.commandLists = userData.commandLists.filter(x => x !== list)),
+    {
+      message: `Are you sure you want to delete the list "${list.name}"?`,
+    },
+  );
 }
 
 const dragging = ref(false);
@@ -38,7 +47,7 @@ const draggableOptions = {
 
 <template>
   <ActionBar>
-    <PrunButton primary @click="createNewNote">CREATE NEW</PrunButton>
+    <PrunButton primary @click="createNew">CREATE NEW</PrunButton>
   </ActionBar>
   <table>
     <thead>
@@ -49,24 +58,24 @@ const draggableOptions = {
       </tr>
     </thead>
     <tbody
-      v-draggable="[userData.notes, draggableOptions]"
+      v-draggable="[userData.commandLists, draggableOptions]"
       :class="dragging ? $style.dragging : null">
-      <tr v-for="note in userData.notes" :key="note.id">
+      <tr v-for="list in userData.commandLists" :key="list.id">
         <td>
           <span :class="[grip.grip, fa.solid, $style.grip]">
             {{ '\uf58e' }}
           </span>
-          <PrunLink inline :command="`XIT NOTE ${note.id.substring(0, 8)}`">
-            {{ note.name }}
+          <PrunLink inline :command="`XIT CMDL ${list.id.substring(0, 8)}`">
+            {{ list.name }}
           </PrunLink>
         </td>
         <td>
           <span>
-            {{ note.text.length.toLocaleString() }} character{{ note.text.length !== 1 ? 's' : '' }}
+            {{ list.commands.length }} command{{ list.commands.length !== 1 ? 's' : '' }}
           </span>
         </td>
         <td>
-          <PrunButton danger @click="confirmDelete($event, note)">DELETE</PrunButton>
+          <PrunButton danger @click="confirmDelete($event, list)">DELETE</PrunButton>
         </td>
       </tr>
     </tbody>