+++ src/features/XIT/index.ts
@@ -2,6 +2,7 @@ import './ACT/ACT';
 import './BURN/BURN';
 import './CALC';
 import './CHAT';
+import './CMDL/CMDL';
 import './CMDS';
 import './CONTC/CONTC';
 import './CONTS/CONTS';